Use Short for 16 bit values

+++ app/src/main/java/com/example/blessed3/BloodPressureMeasurementStatus.kt
@@ -1,6 +1,6 @@
 package com.example.blessed3
 
-class BloodPressureMeasurementStatus internal constructor(measurementStatus: UInt) {
+class BloodPressureMeasurementStatus internal constructor(measurementStatus: UShort) {
     /**
      * Body Movement Detected
      */
